Create basic skeleton with contexts for search page
