+++ client/src/app/pages/package-list/package-table.tsx
@@ -1,0 +1,152 @@
+import React from "react";
+import { NavLink } from "react-router-dom";
+
+import {
+  DescriptionList,
+  DescriptionListDescription,
+  DescriptionListGroup,
+  DescriptionListTerm,
+} from "@patternfly/react-core";
+import {
+  ExpandableRowContent,
+  Table,
+  Tbody,
+  Td,
+  Th,
+  Thead,
+  Tr,
+} from "@patternfly/react-table";
+
+import { PackageQualifiers } from "@app/components/PackageQualifiers";
+import { SimplePagination } from "@app/components/SimplePagination";
+import {
+  ConditionalTableBody,
+  TableHeaderContentWithControls,
+  TableRowContentWithControls,
+} from "@app/components/TableControls";
+
+import { PackageSearchContext } from "./package-context";
+
+export const PackageTable: React.FC = () => {
+  const { isFetching, fetchError, totalItemCount, tableControls } =
+    React.useContext(PackageSearchContext);
+
+  const {
+    numRenderedColumns,
+    currentPageItems,
+    propHelpers: {
+      paginationProps,
+      tableProps,
+      getThProps,
+      getTrProps,
+      getTdProps,
+    },
+    expansionDerivedState: { isCellExpanded },
+  } = tableControls;
+
+  return (
+    <>
+      <Table {...tableProps} aria-label="Package table">
+        <Thead>
+          <Tr>
+            <TableHeaderContentWithControls {...tableControls}>
+              <Th {...getThProps({ columnKey: "name" })} />
+              <Th {...getThProps({ columnKey: "namespace" })} />
+              <Th {...getThProps({ columnKey: "version" })} />
+              <Th {...getThProps({ columnKey: "type" })} />
+              <Th {...getThProps({ columnKey: "qualifiers" })} />
+            </TableHeaderContentWithControls>
+          </Tr>
+        </Thead>
+        <ConditionalTableBody
+          isLoading={isFetching}
+          isError={!!fetchError}
+          isNoData={totalItemCount === 0}
+          numRenderedColumns={numRenderedColumns}
+        >
+          {currentPageItems.map((item, rowIndex) => {
+            return (
+              <Tbody key={item.uuid}>
+                <Tr {...getTrProps({ item })}>
+                  <TableRowContentWithControls
+                    {...tableControls}
+                    item={item}
+                    rowIndex={rowIndex}
+                  >
+                    <Td width={25} {...getTdProps({ columnKey: "name" })}>
+                      <NavLink
+                        to={`/packages/${encodeURIComponent(item.uuid)}`}
+                      >
+                        {item.decomposedPurl
+                          ? item.decomposedPurl?.name
+                          : item.purl}
+                      </NavLink>
+                    </Td>
+                    <Td
+                      width={10}
+                      modifier="truncate"
+                      {...getTdProps({ columnKey: "namespace" })}
+                    >
+                      {item.decomposedPurl?.namespace}
+                    </Td>
+                    <Td
+                      width={15}
+                      modifier="truncate"
+                      {...getTdProps({ columnKey: "version" })}
+                    >
+                      {item.decomposedPurl?.version}
+                    </Td>
+                    <Td
+                      width={10}
+                      modifier="truncate"
+                      {...getTdProps({ columnKey: "type" })}
+                    >
+                      {item.decomposedPurl?.type}
+                    </Td>
+                    <Td width={40} {...getTdProps({ columnKey: "qualifiers" })}>
+                      {item.decomposedPurl?.qualifiers && (
+                        <PackageQualifiers
+                          value={item.decomposedPurl?.qualifiers}
+                        />
+                      )}
+                    </Td>
+                  </TableRowContentWithControls>
+                </Tr>
+                {isCellExpanded(item) ? (
+                  <Tr isExpanded>
+                    <Td colSpan={7}>
+                      <ExpandableRowContent>
+                        <div className="pf-v5-u-m-md">
+                          <DescriptionList>
+                            <DescriptionListGroup>
+                              <DescriptionListTerm>Purl</DescriptionListTerm>
+                              <DescriptionListDescription>
+                                {item.purl}
+                              </DescriptionListDescription>
+                            </DescriptionListGroup>
+                            <DescriptionListGroup>
+                              <DescriptionListTerm>Path</DescriptionListTerm>
+                              <DescriptionListDescription>
+                                {item.decomposedPurl?.path}
+                              </DescriptionListDescription>
+                            </DescriptionListGroup>
+                          </DescriptionList>
+                        </div>
+                      </ExpandableRowContent>
+                    </Td>
+                  </Tr>
+                ) : null}
+              </Tbody>
+            );
+          })}
+        </ConditionalTableBody>
+      </Table>
+      <SimplePagination
+        idPrefix="package-table"
+        isTop={false}
+        isCompact
+        paginationProps={paginationProps}
+      />
+    </>
+  );
+};
