+++ client/src/app/pages/sbom-list/sbom-context.tsx
@@ -1,0 +1,100 @@
+import React from "react";
+
+import { AxiosError } from "axios";
+
+import { SbomSummary } from "@app/client";
+import { FilterType } from "@app/components/FilterToolbar";
+import { TablePersistenceKeyPrefixes } from "@app/Constants";
+import {
+  getHubRequestParams,
+  ITableControls,
+  useTableControlProps,
+  useTableControlState,
+} from "@app/hooks/table-controls";
+import { useSelectionState } from "@app/hooks/useSelectionState";
+import { useFetchSBOMs } from "@app/queries/sboms";
+
+interface ISbomSearchContext {
+  tableControls: ITableControls<
+    SbomSummary,
+    "name" | "published" | "labels" | "packages",
+    "published",
+    "",
+    string
+  >;
+
+  totalItemCount: number;
+  isFetching: boolean;
+  fetchError: AxiosError;
+}
+
+const contextDefaultValue = {} as ISbomSearchContext;
+
+export const SbomSearchContext =
+  React.createContext<ISbomSearchContext>(contextDefaultValue);
+
+interface ISbomProvider {
+  children: React.ReactNode;
+}
+
+export const SbomSearchProvider: React.FunctionComponent<ISbomProvider> = ({
+  children,
+}) => {
+  const tableControlState = useTableControlState({
+    tableName: "sbom",
+    persistenceKeyPrefix: TablePersistenceKeyPrefixes.sboms,
+    columnNames: {
+      name: "Name",
+      published: "Published",
+      labels: "Labels",
+      packages: "Packages",
+    },
+    isPaginationEnabled: true,
+    isSortEnabled: true,
+    sortableColumns: ["published"],
+    isFilterEnabled: true,
+    filterCategories: [
+      {
+        categoryKey: "",
+        title: "Filter text",
+        placeholderText: "Search",
+        type: FilterType.search,
+      },
+    ],
+    isExpansionEnabled: true,
+    expandableVariant: "compound",
+  });
+
+  const {
+    result: { data: advisories, total: totalItemCount },
+    isFetching,
+    fetchError,
+  } = useFetchSBOMs(
+    getHubRequestParams({
+      ...tableControlState,
+      hubSortFieldKeys: {
+        published: "published",
+      },
+    })
+  );
+
+  const tableControls = useTableControlProps({
+    ...tableControlState,
+    idProperty: "id",
+    currentPageItems: advisories,
+    totalItemCount,
+    isLoading: isFetching,
+    selectionState: useSelectionState({
+      items: advisories,
+      isEqual: (a, b) => a.id === b.id,
+    }),
+  });
+
+  return (
+    <SbomSearchContext.Provider
+      value={{ totalItemCount, isFetching, fetchError, tableControls }}
+    >
+      {children}
+    </SbomSearchContext.Provider>
+  );
+};
