+++ client/src/app/pages/sbom-list/sbom-toolbar.tsx
@@ -1,0 +1,38 @@
+import React from "react";
+
+import { Toolbar, ToolbarContent, ToolbarItem } from "@patternfly/react-core";
+
+import { FilterToolbar } from "@app/components/FilterToolbar";
+import { SimplePagination } from "@app/components/SimplePagination";
+
+import { SbomSearchContext } from "./sbom-context";
+
+interface ISbomToolbar {}
+
+export const SbomToolbar: React.FC<ISbomToolbar> = ({}) => {
+  const { tableControls } = React.useContext(SbomSearchContext);
+
+  const {
+    propHelpers: {
+      toolbarProps,
+      filterToolbarProps,
+      paginationToolbarItemProps,
+      paginationProps,
+    },
+  } = tableControls;
+
+  return (
+    <Toolbar {...toolbarProps}>
+      <ToolbarContent>
+        <FilterToolbar {...filterToolbarProps} />
+        <ToolbarItem {...paginationToolbarItemProps}>
+          <SimplePagination
+            idPrefix="sbom-table"
+            isTop
+            paginationProps={paginationProps}
+          />
+        </ToolbarItem>
+      </ToolbarContent>
+    </Toolbar>
+  );
+};
