+++ client/src/app/pages/search/index.ts
@@ -1,1 +1,1 @@
-export { Search as default } from "./search";
+export { SearchPage as default } from "./search";
