+++ client/src/app/pages/vulnerability-list/vulnerability-context.tsx
@@ -1,0 +1,116 @@
+import React from "react";
+
+import { AxiosError } from "axios";
+
+import { VulnerabilitySummary } from "@app/client";
+import { FilterType } from "@app/components/FilterToolbar";
+import { TablePersistenceKeyPrefixes } from "@app/Constants";
+import {
+  getHubRequestParams,
+  ITableControls,
+  useTableControlProps,
+  useTableControlState,
+} from "@app/hooks/table-controls";
+import { useSelectionState } from "@app/hooks/useSelectionState";
+import { useFetchVulnerabilities } from "@app/queries/vulnerabilities";
+
+interface IVulnerabilitySearchContext {
+  tableControls: ITableControls<
+    VulnerabilitySummary,
+    "packages" | "identifier" | "title" | "severity" | "sboms" | "advisories",
+    never,
+    "" | "average_severity",
+    string
+  >;
+
+  totalItemCount: number;
+  isFetching: boolean;
+  fetchError: AxiosError;
+}
+
+const contextDefaultValue = {} as IVulnerabilitySearchContext;
+
+export const VulnerabilitySearchContext =
+  React.createContext<IVulnerabilitySearchContext>(contextDefaultValue);
+
+interface IVulnerabilityProvider {
+  children: React.ReactNode;
+}
+
+export const VulnerabilitySearchProvider: React.FunctionComponent<
+  IVulnerabilityProvider
+> = ({ children }) => {
+  const tableControlState = useTableControlState({
+    tableName: "vulnerability",
+    persistenceKeyPrefix: TablePersistenceKeyPrefixes.vulnerabilities,
+    columnNames: {
+      identifier: "Identifier",
+      title: "Title",
+      severity: "Severity",
+      packages: "Packages",
+      sboms: "SBOMs",
+      advisories: "Advisories",
+    },
+    isPaginationEnabled: true,
+    isSortEnabled: true,
+    sortableColumns: [],
+    isFilterEnabled: true,
+    filterCategories: [
+      {
+        categoryKey: "",
+        title: "Filter text",
+        placeholderText: "Search",
+        type: FilterType.search,
+      },
+      {
+        categoryKey: "average_severity",
+        title: "Severity",
+        placeholderText: "Severity",
+        type: FilterType.multiselect,
+        selectOptions: [
+          { value: "none", label: "None" },
+          { value: "low", label: "Low" },
+          { value: "medium", label: "Medium" },
+          { value: "high", label: "High" },
+          { value: "critical", label: "Critical" },
+        ],
+      },
+    ],
+    isExpansionEnabled: true,
+    expandableVariant: "compound",
+  });
+
+  const {
+    result: { data: vulnerabilities, total: totalItemCount },
+    isFetching,
+    fetchError,
+  } = useFetchVulnerabilities(
+    getHubRequestParams({
+      ...tableControlState,
+      hubSortFieldKeys: {
+        identifier: "identifier",
+        severity: "average_score",
+      },
+    })
+  );
+
+  const tableControls = useTableControlProps({
+    ...tableControlState,
+    idProperty: "identifier",
+    currentPageItems: vulnerabilities,
+    totalItemCount,
+    isLoading: isFetching,
+    selectionState: useSelectionState({
+      items: vulnerabilities,
+      isEqual: (a, b) => a.identifier === b.identifier,
+    }),
+  });
+
+  return (
+    <VulnerabilitySearchContext.Provider
+      value={{ totalItemCount, isFetching, fetchError, tableControls }}
+    >
+      {children}
+    </VulnerabilitySearchContext.Provider>
+  );
+};
