+++ client/src/app/pages/vulnerability-list/vulnerability-toolbar.tsx
@@ -1,0 +1,38 @@
+import React from "react";
+
+import { Toolbar, ToolbarContent, ToolbarItem } from "@patternfly/react-core";
+
+import { FilterToolbar } from "@app/components/FilterToolbar";
+import { SimplePagination } from "@app/components/SimplePagination";
+
+import { VulnerabilitySearchContext } from "./vulnerability-context";
+
+interface IVulnerabilityToolbar {}
+
+export const VulnerabilityToolbar: React.FC<IVulnerabilityToolbar> = ({}) => {
+  const { tableControls } = React.useContext(VulnerabilitySearchContext);
+
+  const {
+    propHelpers: {
+      toolbarProps,
+      filterToolbarProps,
+      paginationToolbarItemProps,
+      paginationProps,
+    },
+  } = tableControls;
+
+  return (
+    <Toolbar {...toolbarProps}>
+      <ToolbarContent>
+        <FilterToolbar {...filterToolbarProps} />
+        <ToolbarItem {...paginationToolbarItemProps}>
+          <SimplePagination
+            idPrefix="vulnerability-table"
+            isTop
+            paginationProps={paginationProps}
+          />
+        </ToolbarItem>
+      </ToolbarContent>
+    </Toolbar>
+  );
+};
